Migration from SWAT-Research (#11)
Brings 793 commits of development work including:

- Symbolic execution enhancements and new dispatcher architecture
- Spring framework instrumentation support
- Symbolic array handling and annotation-based symbolization
- Enhanced constraint solving with caching
- New analysis modules and coverage tracking
- HTTP driver and SV-COMP integration
- Z3 solver library updates (4.12.5 -> 4.15.4)
- Additional target applications and test cases

+++ symbolic-executor/src/main/java/de/uzl/its/swat/symbolic/trace/dto/BranchDTO.java
@@ -2,7 +2,7 @@ package de.uzl.its.swat.symbolic.trace.dto;
 
 public class BranchDTO {
     @SuppressWarnings("unused")
-    private final int iid;
+    private final long iid;
 
     @SuppressWarnings("unused")
     private String constraint;
@@ -16,14 +16,14 @@ public class BranchDTO {
     @SuppressWarnings("unused")
     private String inst;
 
-    public BranchDTO(int iid, String constraint, boolean branched) {
+    public BranchDTO(long iid, String constraint, boolean branched) {
         this.iid = iid;
         this.constraint = constraint;
         this.branched = branched;
         this.type = "Branch";
     }
 
-    public BranchDTO(int iid, String inst) {
+    public BranchDTO(long iid, String inst) {
         this.iid = iid;
         this.inst = inst;
         this.type = "Special";
