+++ symbolic-executor/src/main/java/de/uzl/its/swat/symbolic/value/reference/LambdaPlaceHolder.java
@@ -3,11 +3,11 @@ package de.uzl.its.swat.symbolic.value.reference;
 import de.uzl.its.swat.symbolic.value.Value;
 
 public class LambdaPlaceHolder extends Value {
-    public int getKey() {
+    public long getKey() {
         return key;
     }
 
-    private final int key;
+    private final long key;
 
     public int getParentAddress() {
         return parentAddress;
@@ -16,7 +16,7 @@ public class LambdaPlaceHolder extends Value {
     // parentAddress is -1 if the lambda function is static
     private final int parentAddress;
 
-    public LambdaPlaceHolder(int key, int parentAddress) {
+    public LambdaPlaceHolder(long key, int parentAddress) {
         this.key = key;
         this.parentAddress = parentAddress;
     }
